+++ src/main/java/roomescape/model/Reservation.java
@@ -1,6 +1,7 @@
 package roomescape.model;
 
 import roomescape.model.member.Member;
+import roomescape.model.theme.Theme;
 import roomescape.service.dto.ReservationDto;
 
 import java.time.LocalDate;
