+++ src/main/java/roomescape/model/theme/Theme.java
@@ -1,4 +1,4 @@
-package roomescape.model;
+package roomescape.model.theme;
 
 import roomescape.service.dto.ThemeDto;
