+++ src/main/java/roomescape/service/ReservationService.java
@@ -7,7 +7,7 @@ import roomescape.exception.NotFoundException;
 import roomescape.model.member.Member;
 import roomescape.model.Reservation;
 import roomescape.model.ReservationTime;
-import roomescape.model.Theme;
+import roomescape.model.theme.Theme;
 import roomescape.repository.ReservationRepository;
 import roomescape.service.dto.ReservationDto;
 import roomescape.service.dto.ReservationTimeInfoDto;
