+++ src/test/java/roomescape/model/ReservationTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import roomescape.model.member.Member;
 import roomescape.model.member.Role;
+import roomescape.model.theme.Theme;
 import roomescape.service.dto.ReservationDto;
 
 import java.time.LocalDate;
