+++ src/test/java/roomescape/model/ThemeTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+import roomescape.model.theme.Theme;
 import roomescape.service.dto.ThemeDto;
 
 import static org.assertj.core.api.Assertions.*;
