+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -9,6 +9,7 @@ import roomescape.exception.NotFoundException;
 import roomescape.model.*;
 import roomescape.model.member.Member;
 import roomescape.model.member.Role;
+import roomescape.model.theme.Theme;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.dao.MemberDao;
 import roomescape.repository.dao.ReservationDao;
